requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.1
idna==3.7
pillow==10.3.0
PyJWT==2.8.0
razorpay==1.4.2
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
